refactor: 整理设置选项顺序和分类
