+++ src/components/settings/index.ts
@@ -1,7 +1,7 @@
 export { AppearanceSection } from './AppearanceSection';
 export { SortableAccentTile } from './SortableAccentTile';
 export { HotkeySection } from './HotkeySection';
-export { MiscSection } from './MiscSection';
+export { GeneralSection } from './GeneralSection';
 export { UpdateSection } from './UpdateSection';
 export { DebugSection } from './DebugSection';
 export { AboutSection } from './AboutSection';
